created an agents folder to host all agent modules moved hf_mixtral_agent to new agents module folder
