+++ example.env
@@ -14,3 +14,5 @@ VECTOR_DATABASE_LOCATION=
 CONVERSATION_COLLECTION_NAME="ConversationMemory"
 
 EMBEDDING_MODEL="sentence-transformers/all-MiniLM-L6-v2"
+
+SOURCES_CACHE = "innovation_pathfinder_ai/database/sources_cache.sqlite3"
