+++ innovation_pathfinder_ai/agents/ollama_mixtral_agent.py
